# api/docs/swagger.yaml
definitions:
  models.AddPermission:
    properties:
      description:
        example: Can add a new flowers
        type: string
      name:
        example: create_flower
        type: string
    type: object
  models.AddRole:
    properties:
      description:
        example: Manages only orders
        type: string
      name:
        example: Manager
        type: string
    type: object
  models.AddUser:
    properties:
      email:
        example: jumaniyozovibodulla07@gmail.com
        type: string
      language:
        enum:
        - uz
        - ' ru'
        - ' en'
        type: string
      name:
        example: Jumaniyozov Ibodulla
        type: string
      password_hash:
        example: PASSWORD_HASH
        type: string
    type: object
  models.GetAllUsersResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.GetUser'
        type: array
    type: object
  models.GetUser:
    properties:
      created_at:
        type: integer
      email:
        example: jumaniyozovibodulla07@gmail.com
        type: string
      id:
        example: 9c1242b2-b211-4a91-ba6c-a58e903327fd
        type: string
      language:
        enum:
        - uz
        - ' ru'
        - ' en'
        type: string
      name:
        example: Jumaniyozov Ibodulla
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.SeccessRequest:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.UpdatePermission:
    properties:
      description:
        example: Can remove a flower
        type: string
      id:
        example: ccae8606-8be2-4621-9086-4144214b167b
        type: string
      name:
        example: delete_flower
        type: string
    type: object
  models.UpdateRole:
    properties:
      description:
        example: Manages only orders
        type: string
      id:
        example: 88361fd2-b050-4f41-ba25-e262aa7a3113
        type: string
      name:
        example: Manager
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FLOWER SHOP API
  version: "1.0"
paths:
  /api/v1/permission:
    post:
      consumes:
      - application/json
      description: THIS API CREATE AN PERMISSION
      parameters:
      - description: PERMISSION INFO
        in: body
        name: permissions
        required: true
        schema:
          $ref: '#/definitions/models.AddPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: CREATE A PERMISSION
      tags:
      - PERMISSIONS
    put:
      consumes:
      - application/json
      description: THIS API UPDATE THE PERMISSION
      parameters:
      - description: ROLE INFO
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: UPDATE THE PERMISSION
      tags:
      - PERMISSIONS
  /api/v1/permission/{id}:
    delete:
      consumes:
      - application/json
      description: THIS API DELETE THE PERMISSION
      parameters:
      - description: PERMISSION ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: DELETE A PERMISSION
      tags:
      - PERMISSIONS
  /api/v1/permissions:
    get:
      consumes:
      - application/json
      description: THIS API GET ALL PERMISSIONS
      parameters:
      - description: Search BY NAME
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllUsersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: GET ALL PERMISSIONS
      tags:
      - PERMISSIONS
  /api/v1/role:
    post:
      consumes:
      - application/json
      description: THIS API CREATE AN ROLE
      parameters:
      - description: ROLE INFO
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/models.AddRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: CREATE A ROLE
      tags:
      - ROLES
    put:
      consumes:
      - application/json
      description: THIS API UPDATE THE ROLE
      parameters:
      - description: ROLE INFO
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: UPDATE THE ROLE
      tags:
      - ROLES
  /api/v1/role/{id}:
    delete:
      consumes:
      - application/json
      description: THIS API DELETE THE ROLE
      parameters:
      - description: ROLE ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: DELETE A ROLE
      tags:
      - ROLES
  /api/v1/roles:
    get:
      consumes:
      - application/json
      description: THIS API GET ALL ROLES
      parameters:
      - description: Search BY NAME
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllUsersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: GET ALL ROLES
      tags:
      - ROLES
  /api/v1/user:
    post:
      consumes:
      - application/json
      description: THIS API CREATE AN USER
      parameters:
      - description: USER INFO
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/models.AddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: CREATE AN USER
      tags:
      - USERS
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: THIS API DELETE AN USER
      parameters:
      - description: USER ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeccessRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: DELETE AN USER
      tags:
      - USERS
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: THIS API GET ALL USERS
      parameters:
      - description: Search BY NAME
        in: query
        name: search_by_name
        type: string
      - description: Search BY EMAIL
        in: query
        name: search_by_email
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllUsersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: GET ALL USERS
      tags:
      - USERS
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"